math_site/requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2024.10
Django==5.1.4
django-admin-interface==0.29.2
django-colorfield==0.11.0
django-crispy-forms==2.3
django-debug-toolbar==4.4.6
django-extensions==3.2.3
djangorestframework==3.15.2
pillow==11.1.0
python-slugify==8.0.4
sqlparse==0.5.3
text-unidecode==1.3
